requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.6
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-js-asset==2.1.0
django-tinymce==3.6.1
idna==3.4
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
